consolidate chat message components
