+++ src/components/HeroBanner.vue
@@ -4,8 +4,8 @@
       <div class="hero-banner__heading">
         <slot name="heading">
           <h1>
-            One time I said the word
-            <span class="highlight">{{ inputValue}}</span>
+            One time I said a
+            <span class="highlight">bad&nbsp;word</span>
             during an interview&nbsp;🤦.
           </h1>
         </slot>
