+++ src/pages/index.astro
@@ -38,7 +38,7 @@ const initialCurseWord = terminalInputValue.get();
     <p class="text-xl">
       <span class="font-bold">Dummy</span>
       : I value a healthy team culture over all else. I love working with passionate people that love building cool
-      <span class="highlight"><span id="curse-word">{initialCurseWord}</span></span>!
+      <span class="highlight"><span id="curse-word">bad&nbsp;word</span></span>!
     </p>
     <p class="text-md text-center font-bold">*oops</p>
   </PageSection>
